1、添加数组打印函数，打印队列长度改为128 2、HMI屏幕添加车辆号、起始站点、目标站点、当前位置、洁车污车状态、WIFI信号、hmi版本、运行时间的显示 3、删除虚拟进站按键显示及相关逻辑

+++ Core/Inc/main.h
@@ -206,7 +206,7 @@ typedef struct {
     uint32_t dwPlcNum;     // PLC编号（抓包分析该字段都是0x00000001猜测是PLC的编号）
     uint16_t wHeatBeat;    // 心跳信号
     uint16_t wAlm;         // 报警信号
-    uint16_t wCtrl;        // 控制信号
+    uint8_t wCtrl;        // 控制信号
     uint16_t wYear;        // 年
     uint8_t  bMonth;       // 月
     uint8_t  bDay;         // 日
